pyproject.toml
[project]
name = "temperature-semantic-diversity"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.12.0",
    "einops>=0.8.2",
    "inspect-evals[novelty-bench]>=0.3.106",
    "ipykernel>=7.2.0",
    "lovely-tensors>=0.1.21",
    "polars>=1.38.1",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.3",
    "torch>=2.10.0",
    "transformers>=5.2.0",
]
